Exception Handling Code Added

+++ src/main/java/com/ikn/ums/employee/service/EmployeeServiceImpl.java
@@ -19,10 +19,15 @@ import org.springframework.web.client.RestTemplate;
 
 import com.azure.core.credential.AccessToken;
 import com.ikn.ums.employee.VO.DepartmentVO;
+import com.ikn.ums.employee.VO.EmployeeListVO;
 import com.ikn.ums.employee.VO.EmployeeVO;
 import com.ikn.ums.employee.VO.TeamsUserProfileVO;
 import com.ikn.ums.employee.entity.Employee;
-import com.ikn.ums.employee.exception.UMSBusinessException;
+import com.ikn.ums.employee.exception.BusinessException;
+import com.ikn.ums.employee.exception.ErrorCodeMessages;
+import com.ikn.ums.employee.exception.EmptyInputException;
+import com.ikn.ums.employee.exception.EmptyListException;
+import com.ikn.ums.employee.exception.EntityNotFoundException;
 import com.ikn.ums.employee.model.UserProfilesResponseWrapper;
 import com.ikn.ums.employee.repository.EmployeeRepository;
 import com.ikn.ums.employee.utils.InitializeMicrosoftGraph;
@@ -41,11 +46,11 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Autowired
 	private ModelMapper mapper;
-	
+
 	private String accessToken = null;
-	
-	private AccessToken acToken = new AccessToken(this.accessToken,OffsetDateTime.now() );
-	
+
+	private AccessToken acToken = new AccessToken(this.accessToken, OffsetDateTime.now());
+
 	@Autowired
 	private InitializeMicrosoftGraph microsoftGraph;
 
@@ -63,25 +68,19 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public Employee saveEmployee(Employee employee) {
 		log.info("EmployeeService.saveEmployee() ENTERED");
-		try {
-			if ( employee == null) {
-				throw new UMSBusinessException("1022", "Emplyee Object is Null or Empty");
-			}
-			/**
-			 * In case the the handling was not done properly from front end, we need to handle in the Service Layer
-			 */
-			if ( employee.getFirstName().isEmpty() || employee.getFirstName().length() != 0 ) {
-				throw new UMSBusinessException ("1001" , "Please send proper Employee First Name");
-			}
-			return employeeRepository.save(employee);
-		}catch ( IllegalArgumentException e ) {
-			throw new UMSBusinessException("1002","Given Employee is NULL" + e.getMessage()); //Handle exception when employee object is null
-		}catch (Exception e) {
-			throw new UMSBusinessException("1003", "Something went wrong in Service Layer");
+		
+		if (employee == null) {
+			throw new EntityNotFoundException(ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_CODE,
+					ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_MSG);
 		}
-		//TODO: Need to cover all the Business Cases
+		if (employee.getFirstName().isEmpty() || employee.getFirstName().length() != 0) {
+			throw new EmptyInputException(ErrorCodeMessages.ERR_EMP_FIRST_NAME_IS_NULL_CODE,
+					ErrorCodeMessages.ERR_EMP_FIRST_NAME_IS_NULL_MSG);
+		}
+		Employee savedEmployee = employeeRepository.save(employee);
+		return savedEmployee;
 	}
-	
+
 	// fetch user details based on username (email)
 	@Override
 	public EmployeeVO fetchEmployeeDetailsWithDepartment(String email) {
@@ -106,25 +105,37 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	/**
 	 * Retrieve the Employee Details based on the Employee ID
+	 * 
 	 * @param Integer employeeId
 	 * @return EmployeeVO
 	 */
 	public EmployeeVO getEmployeeWithDepartment(Integer employeeId) {
 
-		System.out.println("EmployeeService.getUserWithDepartment() : employeeId : " + employeeId);
-		log.info("EmployeeService.getUserWithDepartment() ENTERED");
-		
+		System.out.println("EmployeeService.getEmployeeWithDepartment() : employeeId : " + employeeId);
+		log.info("EmployeeService.getEmployeeWithDepartment() ENTERED");
+
 		try {
 
-			if ( employeeId != 0) {
-				throw new UMSBusinessException("1010","Employee ID is null");
+			if (employeeId < 0) {
+				System.out.println("EmployeeServiceImpl.getEmployeeWithDepartment() in employee id is null");
+				log.info("EmployeeServiceImpl.getEmployeeWithDepartment() in employee id is null");
+				throw new EmptyInputException ("1010", "Employee ID is null");
 			}
 			// ResponseTemplateVO responseTemplateVO = new ResponseTemplateVO();
 			EmployeeVO employeeVO = new EmployeeVO();
 			Optional<Employee> optEmployee = employeeRepository.findById(employeeId);
+			
+			if (optEmployee.isEmpty())
+				throw new EntityNotFoundException(ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_CODE,
+						ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_MSG);
+
 			Employee employee = optEmployee.get();
-			System.out.println("EmployeeService.getUserWithDepartment() : employee.getDepartmentId() : 0 "
+			System.out.println("EmployeeService.getEmployeeWithDepartment() : employee.getDepartmentId() :  "
 					+ employee.getDepartmentId());
+			
+			
+			
+			
 //				Department department = restTemplate.getForObject("http://localhost:9001/departments/" + employee.getDepartmentId(), Department.class);
 			/**
 			 * There might be multiple instances running over multiple hosts and different
@@ -142,15 +153,15 @@ public class EmployeeServiceImpl implements EmployeeService {
 			employeeVO.setDepartment(department);
 			System.out.println(employeeVO);
 			return employeeVO;
-		}catch ( IllegalArgumentException e) {
-			throw new UMSBusinessException("1008","Employee is null" + e.getMessage());
-		}catch ( NoSuchElementException e ) {
-			throw new UMSBusinessException("1009","Employee is null" + e.getMessage());
+		} catch (IllegalArgumentException e) {
+			throw new BusinessException(ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_CODE,
+					ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_MSG);
+		} catch (NoSuchElementException e) {
+			throw new BusinessException(ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_CODE,
+					ErrorCodeMessages.ERR_EMP_ENTITY_IS_NULL_MSG);
 		}
-		
 	}
 
-	
 //	@Override
 //	public List<Employee> findAllEmployees() {
 //		List<Employee> employeeList = employeeRepository.findAll();
@@ -160,28 +171,25 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public List<Employee> findAllEmployees() {
 		List<Employee> employeesList = null;
-		try {
-			employeesList = employeeRepository.findAll();
-		}catch (Exception e) {
-			throw new UMSBusinessException("1005","Issue in Service Layer when fetching employees");
-		}
-		if (employeesList.isEmpty()) 
-			throw new UMSBusinessException("1004" , "Employees List is Empty");
+		employeesList = employeeRepository.findAll();
+		if ( employeesList == null || employeesList.isEmpty())
+			throw new EmptyListException(ErrorCodeMessages.ERR_EMP_LIST_IS_EMPTY_CODE,
+					ErrorCodeMessages.ERR_EMP_LIST_IS_EMPTY_MSG);
 		return employeesList;
 	}
-	
+
 	@Transactional
 	@SuppressWarnings("rawtypes")
 	@Override
 	public Integer saveAzureUsers() {
 		// get access token from MS teams server , only if it is already null
-				if (this.acToken.isExpired()) {
-					log.info("Access Token expired");
-					 this.acToken = this.microsoftGraph.initializeMicrosoftGraph();
-					 log.info("Access Token Refreshed");
-					 this.accessToken = this.acToken.getToken();
-				}
-		
+		if (this.acToken.isExpired()) {
+			log.info("Access Token expired");
+			this.acToken = this.microsoftGraph.initializeMicrosoftGraph();
+			log.info("Access Token Refreshed");
+			this.accessToken = this.acToken.getToken();
+		}
+
 		// get users from azure active directory
 		String userProfileUrl = "https://graph.microsoft.com/v1.0/users?$filter=accountEnabled eq true and userType eq 'Member'";
 
@@ -200,22 +208,22 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 		// get all user profiles from reponse object
 		List<TeamsUserProfileVO> teamsUserProfilesList = userProfilesResponse.getBody().getValue();
-		
+
 		List<Employee> employeesList = new ArrayList<>();
 		teamsUserProfilesList.forEach(profile -> {
 			Employee e = new Employee();
-			//id will be auto generated
+			// id will be auto generated
 			e.setFirstName(profile.getDisplayName());
 			e.setLastName(profile.getSurname());
 			e.setTeamsUserId(profile.getUserId());
 			e.setEmail(profile.getUserPrincipalName());
 			e.setDesignation(profile.getJobTitle());
 			e.setTwoFactorAuthentication(false);
-			//setting default password, //Test@123 in encrypted format
+			// setting default password, //Test@123 in encrypted format
 			e.setEncryptedPassword("$2a$10$054UvQ85YjjEMnb2Okh9r.qJNDOE9trkRhEjeNE6tdPeeBJNEHZpa");
 			e.setOtpCode(0);
 			e.setDepartmentId(1L);
-			//set default role
+			// set default role
 			e.setUserRole("Team Member");
 			employeesList.add(e);
 		});
@@ -231,48 +239,50 @@ public class EmployeeServiceImpl implements EmployeeService {
 		Employee e = null;
 		TeamsUserProfileVO profile = getAzureUser(azureUserPrincipalName);
 		Employee insertedUser = null;
-		
-		//check for null and convert azure user profile to UMS employee profile
-		if(profile != null) {
+
+		// check for null and convert azure user profile to UMS employee profile
+		if (profile != null) {
 			e = new Employee();
-			//id will be auto generated
+			// id will be auto generated
 			e.setFirstName(profile.getDisplayName());
 			e.setLastName(profile.getSurname());
 			e.setTeamsUserId(profile.getUserId());
 			e.setEmail(profile.getUserPrincipalName());
 			e.setDesignation(profile.getJobTitle());
 			e.setTwoFactorAuthentication(false);
-			//setting default password, //Test@123 in encrypted format
+			// setting default password, //Test@123 in encrypted format
 			e.setEncryptedPassword("$2a$10$054UvQ85YjjEMnb2Okh9r.qJNDOE9trkRhEjeNE6tdPeeBJNEHZpa");
 			e.setOtpCode(0);
 			e.setDepartmentId(1L);
-			//set default role
+			// set default role
 			e.setUserRole("Team Member");
-			
-			//save user
+
+			// save user
 			insertedUser = employeeRepository.save(e);
 		}
-		return "User "+insertedUser.getEmail()+" saved successfully";
+		return "User " + insertedUser.getEmail() + " saved successfully";
 	}
-	
+
 	/**
-	 * get a single user profile from azure active directory based on the user principal name
+	 * get a single user profile from azure active directory based on the user
+	 * principal name
+	 * 
 	 * @param userPrincipalName
 	 * @return
 	 */
 	@SuppressWarnings("rawtypes")
 	private TeamsUserProfileVO getAzureUser(String userPrincipalName) {
-		
+
 		// get access token from MS teams server , only if it is already null
-				if (this.acToken.isExpired()) {
-					log.info("Access Token expired");
-					 this.acToken = this.microsoftGraph.initializeMicrosoftGraph();
-					 log.info("Access Token Refreshed");
-					 this.accessToken = this.acToken.getToken();
-				}
-		
+		if (this.acToken.isExpired()) {
+			log.info("Access Token expired");
+			this.acToken = this.microsoftGraph.initializeMicrosoftGraph();
+			log.info("Access Token Refreshed");
+			this.accessToken = this.acToken.getToken();
+		}
+
 		// get users
-		String userProfileUrl = "https://graph.microsoft.com/v1.0/users/"+userPrincipalName;
+		String userProfileUrl = "https://graph.microsoft.com/v1.0/users/" + userPrincipalName;
 
 		// prepare headers
 		HttpHeaders httpHeaders = new HttpHeaders();
@@ -284,28 +294,31 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 		// prepare the rest template and hit the graph api user end point
 		RestTemplate restTemplate = new RestTemplate();
-		ResponseEntity<TeamsUserProfileVO> userProfileResponse = restTemplate.exchange(userProfileUrl,
-				HttpMethod.GET, httpEntity, TeamsUserProfileVO.class);
+		ResponseEntity<TeamsUserProfileVO> userProfileResponse = restTemplate.exchange(userProfileUrl, HttpMethod.GET,
+				httpEntity, TeamsUserProfileVO.class);
 
 		// get all user profiles from reponse object
 		TeamsUserProfileVO userDto = userProfileResponse.getBody();
 
 		return userDto;
-		
+
 	}
 
 	@Override
 	public Integer searchEmployeeByEmail(String email) {
+		
+		if (email == null || email.isEmpty() )
+			throw new EmptyInputException(ErrorCodeMessages.ERR_EMP_EMAIL_ID_NOT_FOUND_CODE,
+					ErrorCodeMessages.ERR_EMP_EMAIL_ID_NOT_FOUND_MSG);
 		Integer count = employeeRepository.searchEmployeeDetailsByMail(email);
 		return count;
 	}
-	
+
 	@Override
 	public void deleteEmployee(Integer employeeId) {
-		try {
-			employeeRepository.deleteById(employeeId);	
-		}catch (IllegalArgumentException e) {
-			throw new UMSBusinessException("1020","Employee Id is null" + e.getMessage());
-		}
+		if (employeeId ==0 )
+			throw new EmptyInputException(ErrorCodeMessages.ERR_EMP_ID_NOT_FOUND_CODE,
+					ErrorCodeMessages.ERR_EMP_ID_NOT_FOUND_MSG);
+		employeeRepository.deleteById(employeeId);
 	}
 }
